requirements.txt
anthropic
fastapi
uvicorn
psutil
requests
python-dotenv
aiosqlit
groq
